+++ packages/icon/src/lib/Icons.stories.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { Grid } from '@mui/material';
+import { Grid } from '@availity/mui-layout';
 import { FaSvgIconProps } from './FaSvgIcon';
 import { HomeIcon } from './Icons/Common';
 
@@ -25,34 +25,24 @@ export const _HomeIcon: StoryObj<typeof HomeIcon> = {
 export const _Sizes: StoryObj<typeof HomeIcon> = {
   render: () => (
     <Grid container spacing={1}>
-      <Grid item xs={2}>
-        Inherit (default)
-      </Grid>
-      <Grid item xs={10}>
+      <Grid xs={2}>Inherit (default)</Grid>
+      <Grid xs={10}>
         <HomeIcon fontSize="inherit" />
       </Grid>
-      <Grid item xs={2}>
-        XSmall
-      </Grid>
-      <Grid item xs={10}>
+      <Grid xs={2}>XSmall</Grid>
+      <Grid xs={10}>
         <HomeIcon fontSize="xsmall" />
       </Grid>
-      <Grid item xs={2}>
-        Small
-      </Grid>
-      <Grid item xs={10}>
+      <Grid xs={2}>Small</Grid>
+      <Grid xs={10}>
         <HomeIcon fontSize="small" />
       </Grid>
-      <Grid item xs={2}>
-        Medium
-      </Grid>
-      <Grid item xs={10}>
+      <Grid xs={2}>Medium</Grid>
+      <Grid xs={10}>
         <HomeIcon fontSize="medium" />
       </Grid>
-      <Grid item xs={2}>
-        Large
-      </Grid>
-      <Grid item xs={10}>
+      <Grid xs={2}>Large</Grid>
+      <Grid xs={10}>
         <HomeIcon fontSize="large" />
       </Grid>
     </Grid>