+++ packages/paper/src/lib/Paper.tsx
@@ -1,7 +1,7 @@
 import { forwardRef } from 'react';
 import { Paper as MuiPaper, PaperProps as MuiPaperProps } from '@mui/material';
 
-export type PaperProps = Omit<MuiPaperProps, 'classes' | 'style' | 'className' | 'sx' | 'variant'> & {
+export type PaperProps = Omit<MuiPaperProps, 'classes' | 'style' | 'className' | 'variant'> & {
   /**
    * The variant to use.
    * @default 'outlined'