+++ packages/element/CHANGELOG.md
@@ -21,12 +21,12 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ### Dependency Updates
 
-* `mui-progress` updated to version `0.1.0`
-* `theme-provider` updated to version `0.4.3`
+- `mui-progress` updated to version `0.1.0`
+- `theme-provider` updated to version `0.4.3`
 
 ### Features
 
-* **mui-progress:** initial commit ([0012911](https://github.com/Availity/element/commit/00129117d2107d4bdf97a3767119d30ae2edb418))
+- **mui-progress:** initial commit ([0012911](https://github.com/Availity/element/commit/00129117d2107d4bdf97a3767119d30ae2edb418))
 
 ## [0.27.0](https://github.com/Availity/element/compare/@availity/element@0.26.2...@availity/element@0.27.0) (2023-11-09)